# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "e-kot diagramme interactif",
    "version": "0.1.3"
  },
  "servers": [
    {
      "url": "https://aide.e-kot.be"
    }
  ],
  "tags": [
    {
      "name": "diagram",
      "description": "Everything related to diagrams"
    },
    {
      "name": "feedback",
      "description": "Everything related to users feedback"
    }
  ],
  "paths": {
    "/diagrams": {
      "get": {
        "tags": [
          "diagram"
        ],
        "description": "Fetch a preview (first node) of every diagram",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramPreview"
                }
              }
            }
          }
        }
      }
    },
    "/diagram/{id}": {
      "get": {
        "tags": [
          "diagram"
        ],
        "description": "Fetch the requested diagram",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          }
        }
      }
    },
    "/feedback": {
      "get": {
        "tags": [
          "feedback"
        ],
        "description": "Fetch the requested feedbacks",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ids to filter by",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Feedbacks"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity : missing parameter",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "Query parameter 'id' is missing"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "feedback"
        ],
        "description": "Add a new feedback to the db",
        "requestBody": {
          "content": {
            "json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request : the given format is incorrect",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "json format is required"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity : missing property",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "missing property"
                }
              }
            }
          }
        }
      }
    },
    "/feedbacks": {
      "get": {
        "tags": [
          "feedback"
        ],
        "description": "Get information about all feedbacks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbacksOverview"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DiagramPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "answers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Answers"
                }
              }
            }
          }
        }
      },
      "Diagram": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "image": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  }
                }
              },
              "answers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Answers"
                }
              }
            }
          }
        }
      },
      "Answers": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "ref": {
            "type": "integer"
          }
        }
      },
      "Info": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Feedback": {
        "type": "object",
        "required": [
          "reaction"
        ],
        "properties": {
          "reaction": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "diagram": {
            "type": "string"
          }
        }
      },
      "Feedbacks": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "properties": {
              "comment": {
                "type": "string"
              },
              "diagram": {
                "type": "string"
              },
              "date": {
                "type": "string"
              },
              "reaction": {
                "type": "string"
              }
            }
          }
        }
      },
      "FeedbacksOverview": {
        "type": "object",
        "properties": {
          "total": {
            "type": "string"
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}